vaccine: improve tests
